% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vcov_complex}
\alias{vcov_complex}
\title{Calculate complex survey standard errors for lavaan fits estimated with FIML.}
\usage{
vcov_complex(fit, ids = ~1, strata = NULL)
}
\arguments{
\item{fit}{A lavaan Fit object.}

\item{ids}{Clusters specified as a vector or data frame (not a formula!)}

\item{strata}{Strata specified as a vector or data frame (not a formula!)}
}
\value{
The estimated variance-covariance matrix of the paramters, accounting for complex sampling design.
}
\description{
Calculate complex survey standard errors for lavaan fits estimated with FIML.
}
\examples{
library(lavaan)

###### A single group example #######

# European Social Survey Denmark data (SRS)
data(ess.dk, package = "lavaan.survey")

# A saturated model with reciprocal effects from Saris & Gallhofer
dk.model <- "
  socialTrust ~ 1 + systemTrust + fearCrime
  systemTrust ~ 1 + socialTrust + efficacy
  socialTrust ~~ systemTrust
"
lavaan.fit <- lavaan(dk.model, data=ess.dk, auto.var=TRUE, estimator="MLM")
summary(lavaan.fit)
# Use vcov_complex to obtain se's instead of lavaan.survey:
sqrt(diag(vcov_complex(lavaan.fit, ids = ess.dk$intnum)))

# Create random missings and use FIML to estimate the model
set.seed(4897)
ess.dk.mis <- ess.dk
ess.dk.mis[cbind(sample(1:nrow(ess.dk), size=200), sample(3:6, size=200, replace=TRUE))] <- NA
lavaan.fit.mis <- lavaan(dk.model, data=ess.dk.mis, auto.var=TRUE, missing="fiml")
summary(lavaan.fit.mis)
# Use vcov_complex to obtain se's accounting for clustering
sqrt(diag(vcov_complex(lavaan.fit.mis, ids = ess.dk$intnum)))
}

